+++ NEWS.md
@@ -1,5 +1,9 @@
 # leaflet.extras 0.1.2
 
+* Added WMS Legend Control
+
+# leaflet.extras 0.1.2
+
 * Added Leaflet Hash Plugin for bookmarkable/sharable URLs.
 
 # leaflet.extras 0.1.1